+++ src/integrations/types/Salesforce.php
@@ -2,6 +2,7 @@
 
 namespace Solspace\ExpressForms\integrations\types;
 
+use Carbon\Carbon;
 use craft\helpers\UrlHelper;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\RequestException;
@@ -22,6 +23,13 @@ use yii\base\Event;
 
 class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
 {
+    const RESOURCE_LEAD        = 'Lead';
+    const RESOURCE_OPPORTUNITY = 'Opportunity';
+
+    const FIELD_CATEGORY_OPPORTUNITY = 'Opportunity';
+    const FIELD_CATEGORY_ACCOUNT     = 'Account';
+    const FIELD_CATEGORY_CONTACT     = 'Contact';
+
     /** @var string */
     protected $consumerKey;
 
@@ -46,6 +54,12 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
     /** @var string */
     protected $instance;
 
+    /** @var string */
+    protected $closeDate;
+
+    /** @var string */
+    protected $stageName;
+
     /**
      * @return array
      */
@@ -57,6 +71,8 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
             new Setting('Assign Owner?', 'assignOwner', Setting::TYPE_BOOLEAN),
             new Setting('Sandbox mode?', 'sandboxMode', Setting::TYPE_BOOLEAN),
             new Setting('Using custom URL?', 'customUrl', Setting::TYPE_BOOLEAN),
+            new Setting('Close Date (required for Opportunity only)', 'closeDate', Setting::TYPE_TEXT),
+            new Setting('Stage Name (required for Opportunity only)', 'stageName', Setting::TYPE_TEXT),
         ];
     }
 
@@ -357,6 +373,46 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
     }
 
     /**
+     * @return string|null
+     */
+    public function getCloseDate()
+    {
+        return $this->closeDate;
+    }
+
+    /**
+     * @param string $closeDate
+     *
+     * @return Salesforce
+     */
+    public function setCloseDate(string $closeDate = null): Salesforce
+    {
+        $this->closeDate = $closeDate;
+
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getStageName()
+    {
+        return $this->stageName;
+    }
+
+    /**
+     * @param string $stageName
+     *
+     * @return Salesforce
+     */
+    public function setStageName(string $stageName = null): Salesforce
+    {
+        $this->stageName = $stageName;
+
+        return $this;
+    }
+
+    /**
      * @return array
      */
     public function serializeSettings(): array
@@ -370,6 +426,8 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
             'sandboxMode'    => $this->isSandboxMode(),
             'customUrl'      => $this->isCustomUrl(),
             'instance'       => $this->getInstance(),
+            'closeDate'      => $this->getCloseDate(),
+            'stageName'      => $this->getStageName(),
         ];
     }
 
@@ -379,8 +437,8 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
     public function fetchResources(): array
     {
         $resources = [
-            new Resource($this, 'Lead', 'Lead'),
-            // new Resource($this, 'Opportunity', 'Opportunity'),
+            new Resource($this, self::RESOURCE_LEAD, self::RESOURCE_LEAD),
+            new Resource($this, self::RESOURCE_OPPORTUNITY, self::RESOURCE_OPPORTUNITY),
         ];
 
         $event = new FetchResourcesEvent($this, $resources);
@@ -396,10 +454,26 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
      */
     public function fetchResourceFields($resourceId): array
     {
+        if ($resourceId === self::RESOURCE_LEAD) {
+            return $this->fetchFieldsForLeads();
+        }
+
+        if ($resourceId === self::RESOURCE_OPPORTUNITY) {
+            return $this->fetchFieldsForOpportunities();
+        }
+
+        return [];
+    }
+
+    /**
+     * @return array
+     */
+    private function fetchFieldsForLeads(): array
+    {
         $client = $this->generateAuthorizedClient();
 
         try {
-            $response = $client->get($this->getEndpoint("/sobjects/$resourceId/describe"));
+            $response = $client->get($this->getEndpoint('/sobjects/Lead/describe'));
         } catch (RequestException $e) {
             $this->getLogger()->error($e->getMessage(), ['response' => $e->getResponse()]);
 
@@ -425,7 +499,60 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
             $fieldList[] = $fieldObject;
         }
 
-        $event = new FetchResourceFieldsEvent($this, $resourceId, $fieldList);
+        $event = new FetchResourceFieldsEvent($this, self::RESOURCE_LEAD, $fieldList);
+        Event::trigger($this, self::EVENT_FETCH_RESOURCE_FIELDS, $event);
+
+        return $event->getResourceFieldsList();
+    }
+
+    /**
+     * @return array
+     */
+    private function fetchFieldsForOpportunities(): array
+    {
+        $client = $this->generateAuthorizedClient();
+
+        $fieldEndpoints = [
+            self::FIELD_CATEGORY_OPPORTUNITY,
+            self::FIELD_CATEGORY_ACCOUNT,
+            self::FIELD_CATEGORY_CONTACT,
+        ];
+
+        $fieldList = [];
+        foreach ($fieldEndpoints as $category) {
+            try {
+                $response = $client->get($this->getEndpoint("/sobjects/$category/describe"));
+            } catch (RequestException $e) {
+                $this->getLogger()->error($e->getMessage(), ['response' => $e->getResponse()]);
+                continue;
+            }
+
+            $data = \GuzzleHttp\json_decode((string) $response->getBody(), false);
+
+            foreach ($data->fields as $field) {
+                if (!$field->updateable || !empty($field->referenceTo)) {
+                    continue;
+                }
+
+                if (in_array($field->name, ['StageName', 'CloseDate'], true)) {
+                    continue;
+                }
+
+                $fieldObject = new ResourceField(
+                    $field->label,
+                    $category . '___' . $field->name,
+                    $field->type,
+                    !$field->nillable,
+                    (array) $field,
+                    $category
+                );
+
+                $fieldList[] = $fieldObject;
+            }
+
+        }
+
+        $event = new FetchResourceFieldsEvent($this, self::RESOURCE_OPPORTUNITY, $fieldList);
         Event::trigger($this, self::EVENT_FETCH_RESOURCE_FIELDS, $event);
 
         return $event->getResourceFieldsList();
@@ -438,6 +565,25 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
      * @return bool
      */
     public function pushData(IntegrationMappingInterface $mapping, array $postedData = []): bool
+    {
+        if ($mapping->getResourceId() === self::RESOURCE_LEAD) {
+            return $this->pushLeads($mapping, $postedData);
+        }
+
+        if ($mapping->getResourceId() === self::RESOURCE_OPPORTUNITY) {
+            return $this->pushOpportunities($mapping, $postedData);
+        }
+
+        return false;
+    }
+
+    /**
+     * @param IntegrationMappingInterface $mapping
+     * @param array                       $postData
+     *
+     * @return bool
+     */
+    private function pushLeads(IntegrationMappingInterface $mapping, array $postData = []): bool
     {
         $client   = $this->generateAuthorizedClient();
         $endpoint = $this->getEndpoint("/sobjects/{$mapping->getResourceId()}");
@@ -482,6 +628,152 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
             $this->getLogger()->error($responseBody, ['exception' => $e->getMessage()]);
 
             return false;
+        }
+    }
+
+    private function pushOpportunities(IntegrationMappingInterface $mapping, array $postData = []): bool
+    {
+        $mappedValues = [
+            self::FIELD_CATEGORY_OPPORTUNITY => [],
+            self::FIELD_CATEGORY_ACCOUNT     => [],
+            self::FIELD_CATEGORY_CONTACT     => [],
+        ];
+
+        foreach ($mapping->getFieldMappings() as $key => $field) {
+            $resourceField = $mapping->getResourceFields()->get($key);
+            if (null === $resourceField) {
+                continue;
+            }
+
+            if (preg_match('/^\w+___(.*)$/', $key, $matches)) {
+                $key = $matches[1];
+            }
+
+            $mappedValues[$resourceField->getCategory()][$key] = $field->getValueAsString();
+        }
+
+        $event = new IntegrationValueMappingEvent($mappedValues);
+        Event::trigger($this, self::EVENT_AFTER_SET_MAPPING, $event);
+        $mappedValues = $event->getMappedValues();
+
+        $client = $this->generateAuthorizedClient();
+
+        try {
+            $closeDate = new Carbon($this->getCloseDate());
+        } catch (\Exception $e) {
+            $closeDate = new Carbon();
+        }
+
+        $accountMapping     = $mappedValues[self::FIELD_CATEGORY_ACCOUNT] ?? [];
+        $contactMapping     = $mappedValues[self::FIELD_CATEGORY_CONTACT] ?? [];
+        $opportunityMapping = $mappedValues[self::FIELD_CATEGORY_OPPORTUNITY] ?? [];
+
+        $accountName      = $accountMapping['Name'] ?? null;
+        $contactFirstName = $contactMapping['FirstName'] ?? null;
+        $contactLastName  = $contactMapping['LastName'] ?? null;
+        $contactEmail     = $contactMapping['Email'] ?? null;
+        $contactName      = trim("$contactFirstName $contactLastName");
+        if (empty($accountName)) {
+            $accountName            = $contactName;
+            $accountMapping['Name'] = $accountName;
+        }
+
+        $accountRecord = $this->querySingle(
+            "SELECT Id
+                FROM Account
+                WHERE Name = '%s'
+                ORDER BY CreatedDate desc
+                LIMIT 1",
+            [$accountName]
+        );
+
+        $contactRecord = null;
+        if (!empty($contactEmail)) {
+            $contactRecord = $this->querySingle(
+                "SELECT Id
+                FROM Contact
+                WHERE Email = '%s'
+                ORDER BY CreatedDate desc
+                LIMIT 1",
+                [$contactEmail]
+            );
+        }
+
+        if (!$contactRecord) {
+            $contactRecord = $this->querySingle(
+                "SELECT Id
+                FROM Contact
+                WHERE Name = '%s'
+                ORDER BY CreatedDate desc
+                LIMIT 1",
+                [$contactName]
+            );
+        }
+
+        try {
+            if ($accountRecord) {
+                $accountEndpoint = $this->getEndpoint('/sobjects/Account/' . $accountRecord->Id);
+                $response        = $client->patch($accountEndpoint, ['json' => $accountMapping]);
+                $accountId       = $accountRecord->Id;
+
+                Event::trigger($this, self::EVENT_AFTER_RESPONSE, new PushResponseEvent($response));
+            } else {
+                $accountEndpoint = $this->getEndpoint('/sobjects/Account');
+                $accountResponse = $client->post($accountEndpoint, ['json' => $accountMapping]);
+                $accountId       = \GuzzleHttp\json_decode($accountResponse->getBody(), false)->id;
+
+                Event::trigger($this, self::EVENT_AFTER_RESPONSE, new PushResponseEvent($accountResponse));
+            }
+
+            $contactMapping['AccountId'] = $accountId;
+            if ($contactRecord) {
+                $contactEndpoint = $this->getEndpoint('/sobjects/Contact/' . $contactRecord->Id);
+                $response        = $client->patch($contactEndpoint, ['json' => $contactMapping]);
+
+                Event::trigger($this, self::EVENT_AFTER_RESPONSE, new PushResponseEvent($response));
+            } else {
+                $contactEndpoint = $this->getEndpoint('/sobjects/Contact');
+                $response        = $client->post($contactEndpoint, ['json' => $contactMapping]);
+
+                Event::trigger($this, self::EVENT_AFTER_RESPONSE, new PushResponseEvent($response));
+            }
+
+            $opportunityMapping['CloseDate'] = $closeDate->toIso8601ZuluString();
+            $opportunityMapping['AccountId'] = $accountId;
+            $opportunityMapping['StageName'] = $this->getStageName() ?? 'New Stage';
+
+            $response = $client->post(
+                $this->getEndpoint('/sobjects/Opportunity'),
+                ['json' => $opportunityMapping]
+            );
+
+            Event::trigger($this, self::EVENT_AFTER_RESPONSE, new PushResponseEvent($response));
+
+            return $response->getStatusCode() === 201;
+        } catch (RequestException $e) {
+            $exceptionResponse = $e->getResponse();
+            if (!$exceptionResponse) {
+                $this->getLogger()->error($e->getMessage(), ['exception' => $e->getMessage()]);
+
+                throw $e;
+            }
+
+            $responseBody = (string) $exceptionResponse->getBody();
+            $this->getLogger()->error($responseBody, ['exception' => $e->getMessage()]);
+
+            if ($exceptionResponse->getStatusCode() === 400) {
+                $errors = \GuzzleHttp\json_decode((string) $exceptionResponse->getBody(), false);
+
+                if (\is_array($errors)) {
+                    foreach ($errors as $error) {
+                        if (strtoupper($error->errorCode) === 'REQUIRED_FIELD_MISSING') {
+                            return false;
+                        }
+                    }
+                }
+            }
+
+            throw $e;
         }
     }
 
@@ -611,5 +903,78 @@ class Salesforce extends AbstractIntegrationType implements CrmTypeInterface
         }
 
         return $client;
+    }
+
+    /**
+     * @param string $query
+     * @param array  $params
+     *
+     * @return mixed
+     */
+    private function query(string $query, array $params = []): array
+    {
+        $client = $this->generateAuthorizedClient();
+
+        $params = array_map([$this, 'soqlEscape'], $params);
+        $query  = sprintf($query, ...$params);
+
+        try {
+            $response = $client->get(
+                $this->getEndpoint('/query'),
+                [
+                    'query' => [
+                        'q' => $query,
+                    ],
+                ]
+            );
+
+            $result = \GuzzleHttp\json_decode($response->getBody(), false);
+
+            if ($result->totalSize === 0 || !$result->done) {
+                return [];
+            }
+
+            return $result->records;
+        } catch (RequestException $e) {
+            $this->getLogger()->error($e->getMessage(), ['response' => $e->getResponse()]);
+
+            return [];
+        }
+    }
+
+    /**
+     * @param string $query
+     * @param array  $params
+     *
+     * @return mixed|null
+     */
+    private function querySingle(string $query, array $params = [])
+    {
+        $data = $this->query($query, $params);
+
+        if (\count($data) >= 1) {
+            return reset($data);
+        }
+
+        return null;
+    }
+
+    /**
+     * @param string $str
+     *
+     * @return string
+     */
+    private function soqlEscape(string $str = ''): string
+    {
+        $characters  = [
+            '\\',
+            '\'',
+        ];
+        $replacement = [
+            '\\\\',
+            '\\\'',
+        ];
+
+        return str_replace($characters, $replacement, $str);
     }
 }