+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Solspace Express Forms Changelog
 
+## 1.0.0-beta.2 - 2018-04-19
+### Added
+- Added support for Opportunity resource mapping in Salesforce API integration (Pro).
+
+### Fixed
+- Fixed a bug that would cause the Email Notifications CP page to error for some customers.
+
 ## 1.0.0-beta.1 - 2018-04-17
 ### Added
 - Initial release.